+++ artiq/runtime/analyzer.c
@@ -12,7 +12,6 @@ struct analyzer_header {
     unsigned long long int total_byte_count;
     unsigned char overflow_occured;
     unsigned char log_channel;
-    unsigned char dds_channel;
     unsigned char dds_onehot_sel;
 } __attribute__((packed));
 
@@ -72,7 +71,6 @@ void analyzer_start(void)
 
     analyzer_header.overflow_occured = rtio_analyzer_message_encoder_overflow_read();
     analyzer_header.log_channel = CONFIG_RTIO_LOG_CHANNEL;
-    analyzer_header.dds_channel = CONFIG_RTIO_DDS_CHANNEL;
 #ifdef CONFIG_DDS_ONEHOT_SEL
     analyzer_header.dds_onehot_sel = 1;
 #else