+++ examples/master/repository/coredevice_examples/simple/dds_test.py
@@ -6,7 +6,7 @@ class DDSTest(EnvExperiment):
 
     def build(self):
         self.setattr_device("core")
-        self.setattr_device("dds_bus")
+        self.setattr_device("core_dds")
         self.setattr_device("dds0")
         self.setattr_device("dds1")
         self.setattr_device("dds2")
@@ -17,7 +17,7 @@ class DDSTest(EnvExperiment):
 
     @kernel
     def run(self):
-        with self.dds_bus.batch:
+        with self.core_dds.batch:
             self.dds1.set(120*MHz)
             self.dds2.set(200*MHz)
         delay(1*us)