+++ RELEASE_NOTES.rst
@@ -11,4 +11,8 @@ Release notes
   This requires reflashing the runtime and the flash storage filesystem image
   or erase and rewrite its entries.
 * RTIOCollisionError has been renamed to RTIOCollision
- 
+* the new API for DDS batches is:
+    with self.core_dds.batch:
+       ...
+  with core_dds a device of type artiq.coredevice.dds.CoreDDS.
+  The dds_bus device should not be used anymore.